feat: Add Company model and repository interface
The code changes introduce a new `Company` model and a `CompanyRepository` interface in the domain module for the company. This allows for the creation of a new company with an ID and name, as well as the ability to find a company by ID, retrieve all companies, save a company, and delete a company.

Note: The recent user commits and repository commits are not relevant for generating the commit message.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>