feat: Add method to add professional to company
This commit adds a new method `add_professional_to_company` to the `CompanyRepository` trait in the `company_repository.rs` file. This method allows for adding a professional to a company by creating a relationship between them in the Neo4j database. The method takes the `company_id` and `professional_id` as parameters and executes a Cypher query to create the relationship. This functionality is necessary for managing the association between professionals and companies in the application.

Note: The recent user commits and repository commits are not relevant for generating the commit message.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>