chore: Add provenance flag to CI staging workflow
This commit adds the `provenance` flag to the CI staging workflow in order to disable the provenance feature during the build process. This change is necessary to improve the build performance and reduce resource usage.

Note: The recent user commits and repository commits are not relevant for generating the commit message.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>