chore: Add Helm chart for Kubernetes
This commit adds the necessary files and templates to create a Helm chart for deploying the Ferris Company application on Kubernetes. The chart includes the Chart.yaml file with basic metadata, the .helmignore file to specify files to ignore during packaging, and templates for the Service, ServiceAccount, HorizontalPodAutoscaler, and test connection Pod. This Helm chart will make it easier to deploy and manage the application on Kubernetes.

Note: The recent user commits and repository commits are not relevant for generating the commit message.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>