chore: Add find_all method to CompanyService
This commit adds the `find_all` method to the `CompanyService` trait. The method retrieves a list of all companies and returns a `Result` containing the list of companies or an error of type `CompanyError`. This method will be implemented in the `CompanyServiceImpl` struct.

Note: The recent user commits and repository commits are not relevant for generating the commit message.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>