feat: Implement find_all method in CompanyRepository
This commit implements the `find_all` method in the `CompanyRepository` trait. The method retrieves all companies from the Neo4j database and returns a `Result` containing the list of companies or a `CompanyError` if an error occurs during the retrieval process.

Note: The recent user commits and repository commits are not relevant for generating the commit message.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>